+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2017-04-13
+### Added
+- Delta time argument to easing function calls
+- Added linear easing function
+
 ## [1.0.3] - 2017-04-13
 ### Fixed
 - Fixed animation acceleration if setting new target value in 'stop' event callback